+++ src/app/server/rfi-pub-sub-engine.interface.ts
@@ -28,6 +28,7 @@ export enum PayloadTypes {
  */
 export interface Payload<T extends PayloadTypes> {
   marshalled_trace?: string;
+  authorized_user?: string;
   type?: T;
 }
 export interface NodeChangePayload extends Payload<PayloadTypes.NODE_CHANGE> {